requirements.txt
django==1.11.24
djangorestframework==3.9.4